:truck: ranking api 파일 이동 #68
