+++ src/app/router.tsx
@@ -12,12 +12,10 @@ import LogIn from '@/pages/logIn';
 import Game from '@/pages/play/game';
 import LobbyPage, { getLobbyInfo } from '@/pages/play/lobby';
 import Mode from '@/pages/play/mode';
-import Ranking from '@/pages/ranking';
+import Ranking, { getRank } from '@/pages/ranking';
 import Result, { getGameResult } from '@/pages/result';
 import Settings from '@/pages/settings';
 import Tutorial from '@/pages/tutorial';
-
-import { getRank } from '@/services/rank';
 
 import { checkToken } from '@/features/auth';
 import { ROUTE } from '@/shared/constants';
