+++ src/pages/ranking/page.tsx
@@ -1,0 +1,40 @@
+import { useLoaderData } from 'react-router-dom';
+
+import * as constants from './constants';
+import * as styles from './index.css';
+
+import RankingItemBox from '@/components/ranking';
+
+import { useUserInfo } from '@/models/user';
+
+const Ranking = () => {
+  const rankList = useLoaderData() as Profile[];
+  const { user: myRank } = useUserInfo();
+
+  if (!myRank) {
+    return null;
+  }
+
+  return (
+    <div className={styles.rankingFullWrapper}>
+      <div className={styles.rankingTitleWrapper}>
+        <div className={styles.rankingTitleText}>{constants.RANKING_TITLE}</div>
+        {myRank.guest && (
+          <div className={styles.guestText}>{constants.GUEST_TEXT}</div>
+        )}
+      </div>
+      <div className={styles.rankingWrapper}>
+        {rankList.map((item) => (
+          <RankingItemBox key={item.rank} player={item} />
+        ))}
+      </div>
+      {myRank && (
+        <div className={styles.myRankingWrapper}>
+          <RankingItemBox player={myRank} guest={myRank.guest} myRank />
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default Ranking;
